requirements.txt
langchain_community
Ipython
openai
requests
camelot-py[cv]
pandas
PyPDF2
python-dotenv
pdfplumber
ollama